apps/backend/pyproject.toml
[tool.poetry]
name = "tramites-ai-backend"
version = "1.0.0"
description = "Backend FastAPI para sistema de IA de trámites gubernamentales"
authors = ["Tramites AI Team"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = "^5.0.1"
openai = "^1.3.0"
stripe = "^7.8.0"
reportlab = "^4.0.7"
pypdf2 = "^3.0.1"
beautifulsoup4 = "^4.12.2"
selenium = "^4.15.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
Pillow = "^10.1.0"
aiofiles = "^23.2.1"
httpx = "^0.25.2"
celery = "^5.3.4"
flower = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
